// apps/backend/src/tasks/enums/task.enum.ts
export enum TaskStatus {
  TODO = 'todo',
  IN_PROGRESS = 'in_progress',
  DONE = 'done',
}

export enum TaskPriority {
  LOW = 'low',
  MEDIUM = 'medium',
  HIGH = 'high',
}

export enum TaskType {
  CALL = 'call',
  MEETING = 'meeting',
  EMAIL = 'email',
  DEADLINE = 'deadline',
  TODO = 'todo',
}